create useComputerMove hook
